% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltares_colors.R
\name{deltares_colors}
\alias{deltares_colors}
\title{Deltares colorscheme generator}
\usage{
deltares_colors(
  palette = "diverging",
  colorscheme = "DeltaresDefault",
  color = "blauw1",
  shade = "light",
  ncolors = 3
)
}
\arguments{
\item{palette}{Indicates the type of palette. Can be either "diverging" or "sequential". Default is "diverging", which will use distinct Deltares colors based on the colorscheme you've provided.
Set this parameter to "sequential" to generate a a sequential scheme based on one of the Deltares colors.
You can define the starting color with the parameter color and you can choose to go to either white or black by setting the parameter shade to "light" or "black" respectively.}

\item{colorscheme}{Deltares divering color schemes. Default is "DeltaresDefault" (n = 6).
Available schemes are: "DeltaresDefault" (n=6), "DeltaresFull" (n=10), "DeltaresBlues" (n=3), "DeltaresBlues_r" (n=3),
"DeltaresGreens" (n=3), "DeltaresGreens_r" (n=3), "DeltaresOranges" (n=3), "DeltaresOranges_r" (n=3).}

\item{color}{The color that is used when palette is set to "sequential. Default is "blauw1". A full list of color choices and their hexcodes is available on https://github.com/KeesWesdorp/theme_deltares
Will create a color scheme with this color to either white or black, depending on your choice with the parameter shade.}

\item{shade}{Parameter that is used when palette is set to "sequential. Can be either "light" or "dark". Default is "light".
Will create a color scheme to either "white"(#FFFFFF) if you set this to "light" or "black" (#000000) if you set it to "dark"}

\item{ncolors}{Number of colors that are generated when palette is "sequential". Default is 3.}
}
\value{
A character vector.
}
\description{
Deltares colorscheme generator
}
\examples{



deltares_colors() # default color scheme with 6 Deltares colours
deltares_colors(colorscheme = "DeltaresFull") # all Deltares colours (n = 10)
deltares_colors(palette = "sequential", ncolors = 10) # creates sequential color scheme with 10 colors starting with default color "blauw1" (#080c80) all the way to white (#FFFFFF)
deltares_colors(palette = "sequential", color = "blauw3", shade = "dark", ncolors = 8)  # creates sequential scheme from blauw3(#0ebbf0) to black (#000000)

# Use with ggplot:

# create dataset
mtcars2 <- within(mtcars, {
  vs <- factor(vs, labels = c("V-shaped", "Straight"))
  am <- factor(am, labels = c("Automatic", "Manual"))
  cyl  <- factor(cyl)
  carb <- factor(carb)
})

# create ggplot
 p1 <- ggplot2::ggplot(mtcars2) +
 ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = carb)) +
  ggplot2::labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "(1973-74)",
    caption = "Data from the 1974 Motor Trend US magazine.",
    tag = "Figure 1",
    x = "Weight (1000 lbs)",
    y = "Fuel economy (mpg)",
    colour = "Gears"
  )

p1 + scale_colour_manual(values=c(deltares_colors())) # the default Deltares colourscheme

}
