% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_deltares.R
\name{theme_deltares}
\alias{theme_deltares}
\title{Deltares plot theme function}
\usage{
theme_deltares(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Deltares plot theme function
}
\examples{
mtcars2 <- within(mtcars, {
  vs <- factor(vs, labels = c("V-shaped", "Straight"))
  am <- factor(am, labels = c("Automatic", "Manual"))
  cyl  <- factor(cyl)
  gear <- factor(gear)
})



p1 <- ggplot2::ggplot(mtcars2) +
 ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = gear)) +
  ggplot2::labs(
    title = "Fuel economy declines as weight increases",
    subtitle = "(1973-74)",
    caption = "Data from the 1974 Motor Trend US magazine.",
    tag = "Figure 1",
    x = "Weight (1000 lbs)",
    y = "Fuel economy (mpg)",
    colour = "Gears"
  )

p1 + theme_deltares() # the default

p2 <- p1 + ggplot2::facet_grid(vs ~ am)
p2 + theme_deltares()
}
